chore: rename token files + move client + tokens + strava to folders
